requirements.txt
aiohttp==3.6.2
astroid==2.3.3
async-timeout==3.0.1
attrs==19.3.0
certifi==2019.11.28
chardet==3.0.4
idna==2.8
idna-ssl==1.1.0
isort==4.3.21
lazy-object-proxy==1.4.3
mccabe==0.6.1
multidict==4.7.4
pkg-resources==0.0.0
pylint==2.4.4
python-dotenv==0.10.3
requests==2.22.0
six==1.13.0
slackclient==2.5.0
typed-ast==1.4.1
typing-extensions==3.7.4.1
urllib3==1.25.8
wrapt==1.11.2
yarl==1.4.2
